Added markdown for special icons to appear in the icon bar: - links - question mark to indicate something you want to look up or do research on - idea light bulb - checkbox - checked checkbox
Made some stylistic changes in terms of default icons in the icon bar.  The dash will appear when there is text.  A centered dot will appear for a blank line.

Put in logic in mmd.js so it handles asterisk at the beginning, but also handles multiple asterisks at the beginning of a line for emphasis and italic.

Fixed the backspace bug.  Now it behaves as it should.