That was a harder than it should have been.
One stupid hack that I really hate is that when you backspace and it
creates a new range it (for some reason I cannot fathom) it deletes that
last character.  I made it so it adds the last character back so when it
deletes that last character it is still the same string.  I really hate
myself for this hack.  But, I'm tired.  It works.  I haven't found any
side-effects yet.   Yet!!!